Implemented complete exception handling for CRUD operations

+++ Java Assignments/session 2/user-management-system/src/main/java/com/example/usermanagementsystem/service/UserService.java
@@ -3,6 +3,7 @@ package com.example.usermanagementsystem.service;
 import com.example.usermanagementsystem.model.User;
 import com.example.usermanagementsystem.repository.UserRepository;
 import com.example.usermanagementsystem.exception.UserNotFoundException;
+import com.example.usermanagementsystem.exception.UserAlreadyExistsException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,6 +23,7 @@ public class UserService {
         return repo.getAllUsers();
     }
 
+    
     public User getUserById(int id) {
         User user = repo.getUserById(id);
 
@@ -31,18 +33,38 @@ public class UserService {
     return user;
     }
 
+
     public String addUser(User user) {
-        repo.addUser(user);
-        return "User added successfully";
+        User existingUser = repo.getUserById(user.getId());
+
+    if (existingUser != null) {
+        throw new UserAlreadyExistsException("User with this ID already exists");
     }
+    repo.addUser(user);
+    return "User added successfully";
+    }
+
+
 
     public String deleteUser(int id) {
-        repo.deleteUser(id);
-        return "User deleted successfully";
+        User user = repo.getUserById(id);
+
+    if (user == null) {
+        throw new UserNotFoundException("User not found with id: " + id);
+    }
+    repo.deleteUser(id);
+    return "User deleted successfully";
     }
 
+
+
     public String updateUser(int id, String name, String email) {
-        repo.updateUser(id, name, email);
-        return "User updated successfully";
+        User user = repo.getUserById(id);
+
+    if (user == null) {
+        throw new UserNotFoundException("User not found with id: " + id);
+    }
+    repo.updateUser(id, name, email);
+    return "User updated successfully";
     }
 }
